+++ src/main/java/io/github/astrapi69/collection/array/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides utility classes for array objects
+ */
+package io.github.astrapi69.collection.array;