+++ src/main/java/io/github/astrapi69/collection/enumeration/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package provides an adapter class for {@link java.util.Set} and the given
+ * {@link java.util.Enumeration} objects
+ */
+package io.github.astrapi69.collection.enumeration;