+++ src/main/java/io/github/astrapi69/collection/list/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides utility class for list objects
+ */
+package io.github.astrapi69.collection.list;