+++ src/main/java/io/github/astrapi69/collection/modification/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides helper classes for tracking modifications in collections
+ */
+package io.github.astrapi69.collection.modification;