+++ src/main/java/io/github/astrapi69/collection/properties/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package provides a utility class for properties objects and properties implementations for
+ * sorting
+ */
+package io.github.astrapi69.collection.properties;